robot-companion/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
python-engineio>=4.9.0
python-socketio>=5.11.0
opencv-python-headless>=4.9.0
openai>=1.0.0
python-dotenv>=1.0.0
